Creating the projection processing, altering the storage of the Feature to contain Tuples

+++ VectorTileRendering/TileFactory/TileFactory/DataPipeline/ADataPipelineContext.cs
@@ -11,7 +11,7 @@ namespace TileFactory.DataPipeline
         public bool SolidChildren { get; set; }
         public double Tolerance { get; set; }
         public double Extent { get; set; }
-        public int Buffer { get; set; }
+        public double Buffer { get; set; }
         public int LogLevel { get; set; }
 
         public IEnumerable<TileFactory.Feature> TileFeatures { get; set; }
